requirements.txt
django==4.2.26
djangorestframework==3.16.1
django-cors-headers==4.9.0
